add link of notification if path is enter in paraterer of newNotification si la quantity dun produit est inferieur a 1, alors il n'est pas inclut dans la commande probleme de notification envoyé au client resolut maintenant la valeur de chaque produit present dans le panier est afficher dans l'input, et au dessu de l'image du produit dans la homePage ajout du message flash d'annulation de commande passage du type integer a float pour le total des commandes dans l'entité Orders Ajout d'une fonction show dans le PublicOrderController.php

+++ src/Entity/Orders.php
@@ -32,7 +32,7 @@ class Orders
     private $created_at;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $total;
 
@@ -81,12 +81,12 @@ class Orders
         return $this;
     }
 
-    public function getTotal(): ?int
+    public function getTotal(): ?float
     {
         return $this->total;
     }
 
-    public function setTotal(int $total): self
+    public function setTotal(float $total): self
     {
         $this->total = $total;
